(System)Verilog: escaped identifiers (LRM 5.6.1)
Add support for escaped identifiers in Verilog and SystemVerilog:
`\` + zero or more non-whitespace characters + whitespace.

Note that the `\` itself isn't part of the identifier, and that `\foo`
is the same as just `foo` (unlike in VHDL), but that identifiers
identical to keywords such as `\begin` are also allowed.

This definition also theoretically allows an empty identifier, `\`,
which ctags doesn't support.
In that case, leave the `\` as part of the name.
Since this would create ambiguity with an identifier declared as `\\`,
leave the `\` if the name starts with `\\` as well
(so `\foo` becomes `foo` but `\\foo` stays as `\\foo`).

Escaped identifiers are defined in the standard in section
3.7.1 (Verilog) / 5.6.1 (SystemVerilog).